requirements.txt
pandas>=1.3.0
pyyaml>=5.4.0
asyncpg>=0.25.0
boto3>=1.20.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-docker>=2.0.0
docker>=6.0.0
# New dependencies for scheduler and dashboard
redis>=4.0.0
croniter>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0